+++ CHANGELOG.md
@@ -14,3 +14,4 @@
   - style(Asset, Domain, Port): Add tooltips for buttons
   - feat(Port): Experimenting cmd line args
   - feat(Dir): Add basic logic of directory bruteforcing
+  - feat(Finger): Add basic logic of application fingerprinting